requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.7
opencv-python>=4.8.0.76
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0.1
pandas>=2.0.0
scikit-learn>=1.3.0
